extract into gravity error helper;
